refac: Gemini domain을 Habit domain에 병합
