+++ src/main/java/com/wave/wavi/habit/controller/HabitController.java
@@ -2,8 +2,10 @@ package com.wave.wavi.habit.controller;
 
 import com.wave.wavi.common.ResponseDto;
 import com.wave.wavi.config.jwt.JwtUtil;
+import com.wave.wavi.habit.dto.ChatAnalysisRequestDto;
 import com.wave.wavi.habit.dto.HabitRequestDto;
 import com.wave.wavi.habit.dto.HabitResponseDto;
+import com.wave.wavi.habit.service.ChatAnalysisService;
 import com.wave.wavi.habit.service.HabitService;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
@@ -19,6 +21,7 @@ public class HabitController {
 
     private final HabitService habitService;
     private final JwtUtil jwtUtil;
+    private final ChatAnalysisService chatAnalysisService;
 
     // 습관 등록
     @PostMapping("")
@@ -94,6 +97,23 @@ public class HabitController {
         return ResponseDto.builder()
                 .status(HttpStatus.OK.value())
                 .message("습관 상태 갱신 성공")
+                .build();
+    }
+
+    @GetMapping("/chat")
+    public ResponseDto<Object> analyzePrompt(@RequestBody ChatAnalysisRequestDto requestDto) {
+        Object data = chatAnalysisService.analyzePrompt(requestDto);
+        if (data.getClass() == String.class) {
+            return ResponseDto.builder()
+                    .status(HttpStatus.BAD_REQUEST.value())
+                    .message("메시지 분석 결과 정보 부족")
+                    .data(data)
+                    .build();
+        }
+        return ResponseDto.builder()
+                .status(HttpStatus.OK.value())
+                .message("메시지 분석 성공")
+                .data(data)
                 .build();
     }
 }
