+++ src/main/java/com/wave/wavi/habit/service/ChatAnalysisService.java
@@ -1,8 +1,8 @@
-package com.wave.wavi.gemini.service;
+package com.wave.wavi.habit.service;
 
 import com.google.genai.Client;
 import com.google.genai.types.GenerateContentResponse;
-import com.wave.wavi.gemini.dto.ChatAnalysisRequestDto;
+import com.wave.wavi.habit.dto.ChatAnalysisRequestDto;
 import com.wave.wavi.habit.dto.HabitRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.json.simple.JSONObject;
@@ -15,7 +15,7 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class GeminiService {
+public class ChatAnalysisService {
 
     private final Client client;
